Refactor name collision resolution to strategy-based loop
Replace the iteration-number-dependent branching in resolveCollisions
with a clean, uniform loop over an ordered list of named resolution
strategies. The old code used iteration 0 for context suffix, iterations
1-8 for disambiguateName (which itself tried 4 strategies sequentially),
and iteration 9 for numeric fallback.

The new design introduces two function types (collisionGroupStrategy and
disambiguationStrategy) and an ordered strategy list. Each iteration
applies one strategy to all conflicting buckets, then re-buckets. If a
strategy makes no progress, the next one is tried.

Strategies extracted into standalone functions:
- strategyContextSuffix (from resolveWithContextSuffix)
- strategyPerSchemaDisambiguate (wraps 4 sub-strategies from disambiguateName)
- strategyNumericFallback (last-resort safety net)
- tryContentTypeSuffix, tryStatusCodeSuffix, tryParamIndexSuffix,
  tryCompositionTypeSuffix (from disambiguateName)

Also moves all collision resolution code into its own file
(resolve_collisions.go) for better separation of concerns.

Behavioral equivalence verified: go generate produces identical output,
all tests pass.
