+++ experimental/internal/codegen/schemanames.go
@@ -1,7 +1,6 @@
 package codegen
 
 import (
-	"fmt"
 	"strings"
 )
 
@@ -528,210 +527,6 @@ func buildFallbackName(path SchemaPath, converter *NameConverter) string {
 	}
 
 	return converter.ToTypeName(strings.Join(parts, "_"))
-}
-
-// schemaContextSuffix maps a SchemaContext to a disambiguation suffix.
-func schemaContextSuffix(ctx SchemaContext) string {
-	switch ctx {
-	case ContextComponentSchema:
-		return "Schema"
-	case ContextParameter:
-		return "Parameter"
-	case ContextRequestBody:
-		return "Request"
-	case ContextResponse:
-		return "Response"
-	case ContextHeader:
-		return "Header"
-	case ContextCallback:
-		return "Callback"
-	case ContextWebhook:
-		return "Webhook"
-	default:
-		return ""
-	}
-}
-
-// resolveCollisions detects name collisions and makes them unique.
-// Reference schemas are excluded from collision detection because they don't
-// generate types — their names are only used for type resolution lookups.
-//
-// Resolution proceeds in phases:
-//  1. Context suffix: append a suffix derived from the schema's location
-//     (e.g. "Request", "Response"). If exactly one collider lives under
-//     components/schemas it keeps the bare name.
-//  2. Existing disambiguateName logic (content type, status code, composition).
-//  3. Numeric fallback as a last resort.
-func resolveCollisions(schemas []*SchemaDescriptor, candidates map[*SchemaDescriptor]string, converter *NameConverter) {
-	// Filter out reference schemas — they don't generate types so their
-	// short names can safely shadow non-ref names without causing a collision.
-	var nonRefSchemas []*SchemaDescriptor
-	for _, s := range schemas {
-		if s.Ref == "" {
-			nonRefSchemas = append(nonRefSchemas, s)
-		}
-	}
-
-	maxIterations := 10 // Prevent infinite loops
-
-	for iteration := range maxIterations {
-		// Group non-ref schemas by candidate name
-		byName := make(map[string][]*SchemaDescriptor)
-		for _, s := range nonRefSchemas {
-			name := candidates[s]
-			byName[name] = append(byName[name], s)
-		}
-
-		// Check if there are any collisions
-		hasCollisions := false
-		for _, group := range byName {
-			if len(group) > 1 {
-				hasCollisions = true
-				break
-			}
-		}
-
-		if !hasCollisions {
-			return // All names are unique
-		}
-
-		// Resolve collisions
-		for _, group := range byName {
-			if len(group) <= 1 {
-				continue // No collision
-			}
-
-			// On last iteration, just add numeric suffixes
-			if iteration == maxIterations-1 {
-				for i, s := range group {
-					candidates[s] = fmt.Sprintf("%s%d", candidates[s], i+1)
-				}
-				continue
-			}
-
-			// First iteration: try context suffix disambiguation
-			if iteration == 0 {
-				resolveWithContextSuffix(group, candidates)
-				continue
-			}
-
-			// Subsequent iterations: existing disambiguateName logic
-			for i, s := range group {
-				newName := disambiguateName(s, candidates[s], i, converter)
-				candidates[s] = newName
-			}
-		}
-	}
-}
-
-// resolveWithContextSuffix attempts to disambiguate colliding schemas by
-// appending a suffix derived from their path context (e.g. "Request",
-// "Response"). If exactly one member is a component schema, it keeps the
-// bare name and only the others are suffixed.
-func resolveWithContextSuffix(group []*SchemaDescriptor, candidates map[*SchemaDescriptor]string) {
-	// Count how many are from components/schemas
-	var componentSchemaCount int
-	for _, s := range group {
-		ctx, _ := parsePathContext(s.Path)
-		if ctx == ContextComponentSchema {
-			componentSchemaCount++
-		}
-	}
-
-	// If exactly one is from components/schemas, it is "privileged" and keeps
-	// the bare name.
-	privileged := componentSchemaCount == 1
-
-	for _, s := range group {
-		ctx, _ := parsePathContext(s.Path)
-
-		// Privileged component schema keeps the bare name
-		if privileged && ctx == ContextComponentSchema {
-			continue
-		}
-
-		suffix := schemaContextSuffix(ctx)
-		if suffix != "" {
-			name := candidates[s]
-			if !strings.HasSuffix(name, suffix) {
-				candidates[s] = name + suffix
-			}
-		}
-		// If suffix is empty (unknown context), leave unchanged for later
-		// iterations to handle via disambiguateName.
-	}
-}
-
-// disambiguateName adds more context to make a name unique.
-func disambiguateName(s *SchemaDescriptor, currentName string, index int, converter *NameConverter) string {
-	path := s.Path
-
-	// Try to add more path context based on what's in the path
-	// but not already in the name
-
-	// Check for content type differentiation
-	for i, part := range path {
-		if part == "content" && i+1 < len(path) {
-			contentType := path[i+1]
-			var suffix string
-			switch {
-			case strings.Contains(contentType, "json"):
-				suffix = "JSON"
-			case strings.Contains(contentType, "xml"):
-				suffix = "XML"
-			case strings.Contains(contentType, "form"):
-				suffix = "Form"
-			case strings.Contains(contentType, "text"):
-				suffix = "Text"
-			case strings.Contains(contentType, "binary"):
-				suffix = "Binary"
-			default:
-				suffix = converter.ToTypeName(strings.ReplaceAll(contentType, "/", "_"))
-			}
-			// Use Contains since the suffix might be embedded before "Response" or "Request"
-			if !strings.Contains(currentName, suffix) {
-				return currentName + suffix
-			}
-		}
-	}
-
-	// Check for status code differentiation (for responses)
-	for i, part := range path {
-		if part == "responses" && i+1 < len(path) {
-			code := path[i+1]
-			if !strings.Contains(currentName, code) {
-				return currentName + code
-			}
-		}
-	}
-
-	// Check for parameter index differentiation
-	for i, part := range path {
-		if part == "parameters" && i+1 < len(path) {
-			idx := path[i+1]
-			if !strings.HasSuffix(currentName, idx) {
-				return currentName + idx
-			}
-		}
-	}
-
-	// Check for composition type differentiation
-	for i := len(path) - 1; i >= 0; i-- {
-		part := path[i]
-		switch part {
-		case "allOf", "anyOf", "oneOf":
-			suffix := converter.ToTypeName(part)
-			if i+1 < len(path) {
-				suffix += path[i+1] // Add index
-			}
-			if !strings.Contains(currentName, suffix) {
-				return currentName + suffix
-			}
-		}
-	}
-
-	// Last resort: use numeric suffix
-	return fmt.Sprintf("%s%d", currentName, index+1)
 }
 
 // buildOperationRequestName builds a name for a request body using the operationId.
